move out cell styles
