+++ packages/components/src/local/map-control/index.tsx
@@ -5,7 +5,6 @@ import RemoveIcon from '@material-ui/icons/Remove'
 import InfoIcon from '@material-ui/icons/Info'
 
 /* Import proptypes */
-import { CellLabelStyle } from '@serge/config'
 import cx from 'classnames'
 import { DomEvent, LatLngBounds } from 'leaflet'
 import React, { useEffect, useState } from 'react'
@@ -14,11 +13,6 @@ import PropTypes from './types/props'
 import { useMap } from 'react-leaflet-v4'
 import Item from './helpers/item'
 
-interface CellStyleDetails {
-  label: string
-  value: CellLabelStyle
-  active: boolean
-}
 
 export const MapControl: React.FC<PropTypes> = ({
   /* main */
@@ -30,14 +24,11 @@ export const MapControl: React.FC<PropTypes> = ({
   /* zoom */
   showZoom = true,
   zoomStepSize = 0.5,
-  cellLabelCallback,
-  cellLabelType,
   addInfoMarker,
   actionCallback,
   actionItems,
   mapVer = 'v2'
 }) => {
-  const [cellStyles, setCellStyles] = useState<CellStyleDetails[]>([])
   const [originalBounds, setOriginalBounds] = useState<LatLngBounds | undefined>(undefined)
   const localMap = mapVer === 'v4' ? useMap() : map
 
@@ -74,26 +65,7 @@ export const MapControl: React.FC<PropTypes> = ({
     localMap.flyToBounds(originalBounds, { duration: 0.75 })
   }
 
-  /** the forces from props has changed */
-  useEffect(() => {
-    const storeStyle = (label: string, style: CellLabelStyle, current: CellLabelStyle | undefined): CellStyleDetails => {
-      return {
-        label: label,
-        value: style,
-        active: style === current
-      }
-    }
 
-    // collate the cell styles
-    const cellStyleList: CellStyleDetails[] = [
-      storeStyle('Ctr', CellLabelStyle.CTR_LABELS, cellLabelType),
-      storeStyle('H3', CellLabelStyle.H3_LABELS, cellLabelType),
-      storeStyle('L/L', CellLabelStyle.LAT_LON_LABELS, cellLabelType),
-      storeStyle('X-Y', CellLabelStyle.X_Y_LABELS, cellLabelType),
-      storeStyle(' ', CellLabelStyle.BLANK, cellLabelType)
-    ]
-    setCellStyles(cellStyleList)
-  }, [cellLabelType])
 
   if (!localMap) return null
 
@@ -124,18 +96,7 @@ export const MapControl: React.FC<PropTypes> = ({
             </Item>
           </div>
         }
-        {cellLabelCallback && cellStyles.length > 0 && <div className={cx('leaflet-control')}>
-          {cellStyles.map((style: CellStyleDetails): JSX.Element => (
-            <Item
-              contentTheme={style.active ? 'light' : 'dark'}
-              key={`s_${style.value}`}
-              onClick={(): void => { cellLabelCallback && cellLabelCallback(style.value) }}
-              title={'Style cells as ' + style.label}
-            >
-              {style.label}
-            </Item>
-          ))}
-        </div>}
+
         {children}
       </div>
     </div>
